Fix to uniforms and storages
Basically, there was an error in uniforms and storages that made u32 and
i32 to have a wrong representation internally since everything was
converted to Float32Array; this was an original core decision of the
library to avoid complications in the beginning, so basically making
every number an f32 to make things easy. The libraty technically
supports u32 and i32 for a while, and apart from maybe a demo, no issues
were recorded in other demos, but internally the Int numbers were f32s
and the values exploded in size because of the misinterpretation.

Fix removes a few internal assumptions live default converting fo
Float32Array in a few places.

Now it converts appropriately based on the type defined by user on the
JS side.

+++ examples/glb3/structs.js
@@ -9,7 +9,7 @@ struct FragmentCustom {
     @location(4) mouse: vec2f,
     @location(5) normal: vec3f,
     @location(6) world: vec3f,
-    @interpolate(flat) @location(7) id: u32
+    @interpolate(flat) @location(7) id: f32
 }
 
 `;
